GUI stuff and revenue from Humans
Did a lot of stuffs. Enable and disable the build menu after building
has been spawned. Humans generate money that add to each buildings
"bank". Scrolling over a building brings up a context menu with the
building name and a button to collect its money. Money is added to
overall bank and building money is reset to zero